requirements.txt
esprima>=4.0.1
javalang>=0.13.0
pytest>=7.0.0
hypothesis>=6.0.0
google-generativeai>=0.3.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
